+++ examples/cucumber-style-legacy/playwright.config.ts
@@ -4,7 +4,6 @@ import { defineBddConfig, cucumberReporter } from 'playwright-bdd';
 const testDir = defineBddConfig({
   paths: ['features'],
   require: ['steps/*.ts'],
-  importTestFrom: 'steps/fixtures.ts',
 });
 
 export default defineConfig({