+++ examples/cucumber-style/steps/index.ts
@@ -1,15 +1,14 @@
 import { expect } from '@playwright/test';
-import { Given, When, Then } from '@cucumber/cucumber';
-import { CustomWorld } from './world';
+import { Given, When, Then } from './fixtures';
 
-Given<CustomWorld>('I am on home page', async function () {
+Given('I am on home page', async function () {
   await this.openHomePage();
 });
 
-When<CustomWorld>('I click link {string}', async function (name: string) {
+When('I click link {string}', async function (name: string) {
   await this.clickLink(name);
 });
 
-Then<CustomWorld>('I see in title {string}', async function (text: string) {
+Then('I see in title {string}', async function (text: string) {
   await expect(this.page).toHaveTitle(new RegExp(text));
 });