+++ src/gen/index.ts
@@ -23,6 +23,7 @@ import { resovleFeaturePaths } from '../cucumber/resolveFeaturePaths';
 import { loadStepsOwn } from '../cucumber/loadStepsOwn';
 import { relativeToCwd } from '../utils/paths';
 import { BDDConfig } from '../config/types';
+import { appendNewCucumberStyleSteps } from '../steps/cucumberStyle';
 
 export class TestFilesGenerator {
   // all these props are exist
@@ -99,6 +100,7 @@ export class TestFilesGenerator {
       ? await loadStepsOwn(environment.cwd, this.config.steps)
       : await loadSteps(this.runConfiguration, environment);
     await this.loadDecoratorSteps();
+    appendNewCucumberStyleSteps(this.supportCodeLibrary);
     this.logger.log(`Loaded steps: ${this.supportCodeLibrary.stepDefinitions.length}`);
   }