+++ src/steps/createBdd.ts
@@ -1,27 +1,22 @@
 /**
- * Stuff related to writing steps in Playwright-style.
+ * createBdd() definition.
  */
-
-import { DefineStepPattern } from '@cucumber/cucumber/lib/support_code_library_builder/types';
-import { GherkinStepKeyword } from '@cucumber/cucumber/lib/models/gherkin_step_keyword';
 import {
   BuiltInFixtures,
   BuiltInFixturesWorker,
-  FixturesArg,
+  CustomFixtures,
   KeyValue,
   TestTypeCommon,
 } from '../playwright/types';
 import { TestType } from '@playwright/test';
 import { test as baseBddTest } from '../run/bddFixtures';
 import { isTestContainsSubtest } from '../playwright/testTypeImpl';
-import { defineStep } from './defineStep';
 import { exit } from '../utils/exit';
 import { BddWorld } from '../run/bddWorld';
 import { scenarioHookFactory } from '../hooks/scenario';
 import { workerHookFactory } from '../hooks/worker';
-import { fixtureParameterNames } from '../playwright/fixtureParameterNames';
-import { BddAutoInjectFixtures } from '../run/bddFixtures/autoInject';
-import { getLocationByOffset } from '../playwright/getLocationInFile';
+import { CucumberStyleStepFn, cucumberStepCtor } from './cucumberStyle';
+import { PlaywrightStyleStepFn, playwrightStepCtor } from './playwrightStyle';
 
 /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types */
 
@@ -30,47 +25,65 @@ import { getLocationByOffset } from '../playwright/getLocationInFile';
 // todo: https://github.com/vitalets/playwright-bdd/issues/46
 export let hasCustomTest = false;
 
+type CreateBddOptions<WorldFixtureName> = {
+  worldFixture?: WorldFixtureName;
+};
+
+// eslint-disable-next-line max-statements, complexity
 export function createBdd<
   T extends KeyValue = BuiltInFixtures,
   W extends KeyValue = BuiltInFixturesWorker,
-  World extends BddWorld = BddWorld,
->(customTest?: TestType<T, W> | null, _CustomWorld?: new (...args: any[]) => World) {
+  CustomWorld extends BddWorld = BddWorld,
+  WorldFixtureName extends keyof CustomFixtures<T> | undefined = undefined,
+>(
+  customTest?: TestType<T, W> | null,
+  options?: CreateBddOptions<WorldFixtureName> | (new (...args: any[]) => CustomWorld),
+) {
+  // TypeScript does not narrow generic types by control flow
+  // see: https://github.com/microsoft/TypeScript/issues/33912
+  // So, we define return types separately using conditional types
+  type World = WorldFixtureName extends keyof CustomFixtures<T>
+    ? CustomFixtures<T>[WorldFixtureName]
+    : CustomWorld;
+
+  type StepFn = WorldFixtureName extends keyof CustomFixtures<T>
+    ? CucumberStyleStepFn<World>
+    : PlaywrightStyleStepFn<T, W>;
+
+  type StepCtor = WorldFixtureName extends keyof CustomFixtures<T>
+    ? ReturnType<typeof cucumberStepCtor<StepFn>>
+    : ReturnType<typeof playwrightStepCtor<StepFn>>;
+
+  if (options instanceof BddWorld) {
+    // todo: deprecation warning
+  }
+
   if (!hasCustomTest) hasCustomTest = isCustomTest(customTest);
-  const Given = defineStepCtor<T, W>('Given', hasCustomTest);
-  const When = defineStepCtor<T, W>('When', hasCustomTest);
-  const Then = defineStepCtor<T, W>('Then', hasCustomTest);
-  const Step = defineStepCtor<T, W>('Unknown', hasCustomTest);
-  const Before = scenarioHookFactory<T, W, World>('before');
-  const After = scenarioHookFactory<T, W, World>('after');
+
   const BeforeAll = workerHookFactory<W>('beforeAll');
   const AfterAll = workerHookFactory<W>('afterAll');
+
+  // new cucumber-style
+  if (options && 'worldFixture' in options && options.worldFixture) {
+    if (!hasCustomTest) {
+      exit(`When using worldFixture, you should provide custom test to createBdd()`);
+    }
+    const Given = cucumberStepCtor('Given', options.worldFixture) as StepCtor;
+    const When = cucumberStepCtor('When', options.worldFixture) as StepCtor;
+    const Then = cucumberStepCtor('Then', options.worldFixture) as StepCtor;
+    const Step = cucumberStepCtor('Unknown', options.worldFixture) as StepCtor;
+    const Before = scenarioHookFactory<T, W, World>('before', { useWorldFixture: true });
+    const After = scenarioHookFactory<T, W, World>('after', { useWorldFixture: true });
+    return { Given, When, Then, Step, Before, After, BeforeAll, AfterAll };
+  }
+
+  const Given = playwrightStepCtor('Given', hasCustomTest) as StepCtor;
+  const When = playwrightStepCtor('When', hasCustomTest) as StepCtor;
+  const Then = playwrightStepCtor('Then', hasCustomTest) as StepCtor;
+  const Step = playwrightStepCtor('Unknown', hasCustomTest) as StepCtor;
+  const Before = scenarioHookFactory<T, W, World>('before');
+  const After = scenarioHookFactory<T, W, World>('after');
   return { Given, When, Then, Step, Before, After, BeforeAll, AfterAll };
-}
-
-type StepFunctionFixturesArg<T extends KeyValue, W extends KeyValue> = FixturesArg<T, W> &
-  BddAutoInjectFixtures;
-type StepFunction<T extends KeyValue, W extends KeyValue, Args extends any[]> = (
-  fixtures: StepFunctionFixturesArg<T, W>,
-  ...args: Args
-) => unknown;
-
-function defineStepCtor<T extends KeyValue, W extends KeyValue = {}>(
-  keyword: GherkinStepKeyword,
-  hasCustomTest: boolean,
-) {
-  return <Args extends any[]>(pattern: DefineStepPattern, fn: StepFunction<T, W, Args>) => {
-    defineStep({
-      keyword,
-      pattern,
-      fn,
-      hasCustomTest,
-      location: getLocationByOffset(3),
-    });
-    return (fixtures: Partial<StepFunctionFixturesArg<T, W>>, ...args: Args) => {
-      assertStepIsCalledWithRequiredFixtures(pattern, fn, fixtures);
-      return fn(fixtures as StepFunctionFixturesArg<T, W>, ...args);
-    };
-  };
 }
 
 function isCustomTest(customTest?: TestTypeCommon | null) {
@@ -82,28 +95,5 @@ function isCustomTest(customTest?: TestTypeCommon | null) {
 function assertTestHasBddFixtures(customTest: TestTypeCommon) {
   if (!isTestContainsSubtest(customTest, baseBddTest)) {
     exit(`createBdd() should use 'test' extended from "playwright-bdd"`);
-  }
-}
-
-function assertStepIsCalledWithRequiredFixtures<
-  T extends KeyValue,
-  W extends KeyValue,
-  Args extends any[],
->(
-  pattern: DefineStepPattern,
-  fn: StepFunction<T, W, Args>,
-  fixtures: Partial<StepFunctionFixturesArg<T, W>>,
-) {
-  const fixtureNames = fixtureParameterNames(fn);
-  const missingFixtures = fixtureNames.filter(
-    (fixtureName) => !Object.prototype.hasOwnProperty.call(fixtures, fixtureName),
-  );
-  if (missingFixtures.length) {
-    throw new Error(
-      [
-        `Invocation of step "${pattern}" from another step does not pass all required fixtures.`,
-        `Missings fixtures: ${missingFixtures.join(', ')}`,
-      ].join(' '),
-    );
   }
 }