+++ test/call-step-from-step/playwright.config.ts
@@ -4,23 +4,33 @@ import { defineBddConfig } from 'playwright-bdd';
 export default defineConfig({
   projects: [
     {
-      name: 'success',
+      name: 'pw-style-success',
       testDir: defineBddConfig({
-        outputDir: '.features-gen/success',
+        outputDir: '.features-gen/pw-style-success',
         paths: ['features/*.feature'],
-        require: ['steps/index.ts'],
-        importTestFrom: 'steps/fixtures.ts',
+        require: ['steps-pw-style/index.ts'],
+        importTestFrom: 'steps-pw-style/fixtures.ts',
         tags: '@success',
       }),
     },
     {
-      name: 'error',
+      name: 'pw-style-invalid-invocation',
       testDir: defineBddConfig({
-        outputDir: '.features-gen/error',
+        outputDir: '.features-gen/pw-style-invalid-invocation',
         paths: ['features/*.feature'],
-        require: ['steps/index.ts'],
-        importTestFrom: 'steps/fixtures.ts',
+        require: ['steps-pw-style/index.ts'],
+        importTestFrom: 'steps-pw-style/fixtures.ts',
         tags: '@error',
+      }),
+    },
+    {
+      name: 'cucumber-style-success',
+      testDir: defineBddConfig({
+        outputDir: '.features-gen/cucumber-style-success',
+        paths: ['features/*.feature'],
+        require: ['steps-cucumber-style/index.ts'],
+        importTestFrom: 'steps-cucumber-style/fixtures.ts',
+        tags: '@success',
       }),
     },
   ],